project5_agent/requirements.txt
streamlit
boto3
tavily-python
dotenv